attempt tpo compile statically

+++ setup.py
@@ -145,25 +145,45 @@ class RunTwine(Command):
 #
 src = 'src/radical/gtod/gtod.c'
 tgt = 'src/radical/gtod/radical-gtod'
-try:
-    from distutils.ccompiler import new_compiler
 
-    compiler = new_compiler(verbose=1)
-    objs     = compiler.compile(sources=[src])
-    exe      = compiler.link_executable(objs, tgt)
+# try a static build
+if not os.path.exists(tgt):
+    try:
+        compiler = sh_callout('which gcc')[0].decode().strip()
 
-    # test the resulting binary - and if it does not seem to work, replace it
-    # by a shell script which provides a poor-man's version of the C code.
-    now_1 = time.time()
-    out, _, _ = sh_callout(tgt)
-    now_2 = time.time()
-    now   = float(out)
+        if not compiler:
+            compiler = sh_callout('which cc')[0].decode().strip()
 
-    assert(now_1 <= now  )
-    assert(now   <= now_2)
-    assert(now_2  - now_1 <= 1.)
+        assert compiler, 'no compiler found'
 
-except:
+        sh_callout('%s -static -o %s %s' % (compiler, tgt, src))
+
+    except:
+        pass
+
+if not os.path.exists(tgt):
+    try:
+        from distutils.ccompiler import new_compiler
+
+        compiler = new_compiler(verbose=1)
+        objs     = compiler.compile(sources=[src])
+        exe      = compiler.link_executable(objs, tgt)
+
+        # test the resulting binary - and if it does not seem to work, replace it
+        # by a shell script which provides a poor-man's version of the C code.
+        now_1 = time.time()
+        out, _, _ = sh_callout(tgt)
+        now_2 = time.time()
+        now   = float(out)
+
+        assert(now_1 <= now  )
+        assert(now   <= now_2)
+        assert(now_2  - now_1 <= 1.)
+
+    except:
+        pass
+
+if not os.path.exists(tgt):
     # need a replacement
     with open(tgt, 'w') as fout:
         fout.write('''#!/bin/sh
